refactor: extract HookSystem → core/hooks/ (cross-cutting concern)
HookSystem was in core/orchestration/ but is imported by all layers (L0-L4),
creating L0→L3 dependency direction violations.

Move to core/hooks/ as a top-level cross-cutting module:
  core/orchestration/hooks.py       → core/hooks/system.py
  core/orchestration/hook_discovery.py → core/hooks/discovery.py
  core/orchestration/plugins/      → core/hooks/plugins/

core/hooks/__init__.py re-exports HookSystem, HookEvent, HookResult so
all 26 consumers can use `from core.hooks import HookSystem, HookEvent`.

Update core/orchestration/__init__.py docstring: execution control only.
