+++ core/hooks/plugins/notification_hook/hook.yaml
