+++ core/llm/prompt_assembler.py
@@ -14,9 +14,9 @@ import logging
 from dataclasses import dataclass, field
 from typing import Any
 
+from core.hooks import HookEvent, HookSystem
 from core.llm.prompts import _hash_prompt
 from core.llm.skill_registry import SkillDefinition, SkillRegistry
-from core.orchestration.hooks import HookEvent, HookSystem
 
 log = logging.getLogger(__name__)
