+++ core/orchestration/__init__.py
@@ -1,1 +1,1 @@
-"""L4 Orchestration — hook system and event management."""
+"""L3 Orchestration — execution control: TaskGraph, PlanMode, LaneQueue, CoalescingQueue."""
