+++ core/orchestration/task_bridge.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 import logging
 from typing import Any
 
-from core.orchestration.hooks import HookEvent, HookSystem
+from core.hooks import HookEvent, HookSystem
 from core.orchestration.task_system import TaskGraph, TaskStatus
 
 log = logging.getLogger(__name__)
