+++ tests/test_e2e.py
@@ -12,8 +12,8 @@ from core.gateway.auth.cooldown import CooldownTracker
 from core.gateway.auth.profiles import AuthProfile, CredentialType, ProfileStore
 from core.gateway.auth.rotation import ProfileRotator
 from core.graph import compile_graph
+from core.hooks import HookEvent, HookSystem
 from core.llm.router import LLMClientPort
-from core.orchestration.hooks import HookEvent, HookSystem
 from core.state import GeodeState
 from core.tools.analysis import PSMCalculateTool, RunAnalystTool, RunEvaluatorTool
 from core.tools.policy import PolicyChain, ToolPolicy
