+++ tests/test_mcp_lifecycle.py
@@ -16,9 +16,9 @@ import signal
 from typing import Any
 from unittest.mock import MagicMock, patch
 
+from core.hooks import HookEvent
 from core.mcp.manager import MCPServerManager
 from core.mcp.stdio_client import _CLOSE_TIMEOUT_S, StdioMCPClient
-from core.orchestration.hooks import HookEvent
 
 # ---------------------------------------------------------------------------
 # HookEvent tests
