+++ tests/test_notification_hook.py
@@ -7,12 +7,12 @@ from __future__ import annotations
 
 from unittest.mock import MagicMock
 
-from core.mcp.notification_port import NotificationResult, set_notification
-from core.orchestration.hooks import HookEvent, HookSystem
-from core.orchestration.plugins.notification_hook.hook import (
+from core.hooks import HookEvent, HookSystem
+from core.hooks.plugins.notification_hook.hook import (
     _format_message,
     register_notification_hooks,
 )
+from core.mcp.notification_port import NotificationResult, set_notification
 
 # ---------------------------------------------------------------------------
 # Message formatting
