+++ tests/test_project_journal.py
@@ -118,8 +118,8 @@ class TestProjectJournal:
 
 class TestJournalHooks:
     def test_pipeline_end_records_run(self, tmp_path):
+        from core.hooks import HookEvent
         from core.memory.journal_hooks import make_journal_handlers
-        from core.orchestration.hooks import HookEvent
 
         journal = ProjectJournal(tmp_path / "journal")
         handlers = make_journal_handlers(journal)
@@ -148,8 +148,8 @@ class TestJournalHooks:
         assert any("Berserk" in p for p in patterns)
 
     def test_pipeline_error_records(self, tmp_path):
+        from core.hooks import HookEvent
         from core.memory.journal_hooks import make_journal_handlers
-        from core.orchestration.hooks import HookEvent
 
         journal = ProjectJournal(tmp_path / "journal")
         handlers = make_journal_handlers(journal)
@@ -169,8 +169,8 @@ class TestJournalHooks:
         assert runs[0].status == "error"
 
     def test_wrong_event_ignored(self, tmp_path):
+        from core.hooks import HookEvent
         from core.memory.journal_hooks import make_journal_handlers
-        from core.orchestration.hooks import HookEvent
 
         journal = ProjectJournal(tmp_path / "journal")
         handlers = make_journal_handlers(journal)
