+++ tests/test_subagent_announce.py
@@ -27,7 +27,7 @@ from core.agent.sub_agent import (
     drain_announced_results,
 )
 from core.agent.tool_executor import ToolExecutor
-from core.orchestration.hooks import HookEvent, HookSystem
+from core.hooks import HookEvent, HookSystem
 from core.orchestration.isolated_execution import IsolatedRunner
 
 # ---------------------------------------------------------------------------
